+++ docs/versions/v4/v4.md
@@ -1,0 +1,12 @@
+# Fonctionnalités 
+
+- Mot de passe oublié
+- Conservation de session
+- Mise à jour des comptes lors de la connexion
+- Création de l'index dès l'inscription et à la mise à jour du compte
+- Revision des middlewares avec le nouveau service auth2
+- Revision de la langue de l'E-mail de réinitialisation
+- Revision de la langue de base de Laravel, on passe au français
+- Création d'une page dans l'admin pour générer des notifications flash plus facilement (app/http/controller/admin/DevController)
+- Ajout de photo au Modèle d'un utilisateur
+- l'affichage des post ne dépend plus Framework built-in
