+++ docs/versions/v5/news.md
@@ -1,7 +1,6 @@
 # v5.1
 
 - Système de notifications
-- Importation de photo de profil
 - Système de publications
 - Système des gestion des utilisateurs
 - Système de permission
