+++ docs/versions/v5/v5.md
@@ -1,4 +1,4 @@
-# v5.1
+# FAIT
 - Utiliser Dropbox comme CDN pour les photos
 - Revision de la page de profil
 - Amélioration de l'affichage des publications
