+++ resources/views/app/auth/passwords/reset.blade.php
@@ -1,0 +1,74 @@
+@extends('app.page')
+
+@section('extras_style')
+<style media="screen">
+  body{
+    background-color: #f1f3f6
+  }
+</style>
+@endsection
+
+@section('content')
+
+<div class="container mt-5">
+    <div class="row justify-content-center">
+        <div class="col-md-8">
+            <div class="card resac-account-card-support resac-linkedin-shadow p-5">
+                <div class="card-header">Réinitilisationn du mot de passe</div>
+
+                <div class="card-body">
+                    <form method="POST" action="{{ route('app.reinit.password') }}">
+                        @csrf
+
+                        <input type="hidden" name="token" value="{{ $token }}">
+
+                        <div class="form-group row">
+                            <label for="email" class="col-md-4 col-form-label">Adresse E-mail</label>
+
+                            <div class="col-md-8">
+                                <input id="email" type="email" class="form-control @error('email') is-invalid @enderror" name="email" value="{{ $email ?? old('email') }}" required autocomplete="email" autofocus>
+
+                                @error('email')
+                                    <span class="invalid-feedback" role="alert">
+                                        <strong>{{ $message }}</strong>
+                                    </span>
+                                @enderror
+                            </div>
+                        </div>
+
+                        <div class="form-group row">
+                            <label for="password" class="col-md-4 col-form-label">Nouveau mot de passe</label>
+
+                            <div class="col-md-8">
+                                <input id="password" type="password" class="form-control @error('password') is-invalid @enderror" name="password" required autocomplete="new-password">
+
+                                @error('password')
+                                    <span class="invalid-feedback" role="alert">
+                                        <strong>{{ $message }}</strong>
+                                    </span>
+                                @enderror
+                            </div>
+                        </div>
+
+                        <div class="form-group row">
+                            <label for="password-confirm" class="col-md-4 col-form-label">Confirmer le mot de passe</label>
+
+                            <div class="col-md-8">
+                                <input id="password-confirm" type="password" class="form-control" name="password_confirmation" required autocomplete="new-password">
+                            </div>
+                        </div>
+
+                        <div class="form-group row mb-0">
+                            <div class="col-md-8 offset-md-4 text-right">
+                                <button type="submit" class="btn btn-primary">
+                                    Réinitialliser
+                                </button>
+                            </div>
+                        </div>
+                    </form>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
+@endsection
