+++ tests.md
@@ -1,4 +1,5 @@
 ## Test du système d'authentification
+
 - Créer un compte si pas de compte
 - Se connecter
 - Changer le mot de passe du compte depuis la page des paramètres
